refactor: reorganize some files
