+++ graphiql-app/src/components/Footer/components/index.ts
@@ -1,0 +1,3 @@
+import Member from './Member/Member';
+
+export { Member };
