+++ graphiql-app/src/components/Header/components/index.ts
@@ -1,0 +1,4 @@
+import LanguageSwitcher from './LanguageSwitcher/LanguageSwitcher';
+import ThemeSwitcher from './ThemeSwitcher/ThemeSwitcher';
+
+export { LanguageSwitcher, ThemeSwitcher };
