+++ graphiql-app/src/pages/GraphiQL/components/Editor/components/index.ts
@@ -1,0 +1,5 @@
+import Tab from './Tab';
+import { myHighlightStyle, baseTheme } from './editorStyles';
+import { myHighlightStyleL, lightEditorTheme } from './lightEditorTheme';
+
+export { Tab, myHighlightStyle, baseTheme, myHighlightStyleL, lightEditorTheme };
