+++ graphiql-app/src/pages/GraphiQL/components/Editor/components/lightEditorTheme.ts
@@ -1,12 +1,12 @@
 import { EditorView } from '@codemirror/view';
 import { tags } from '@lezer/highlight';
 import { HighlightStyle } from '@codemirror/language';
-import Colors from '../../../../theme/colors';
+import Colors from '../../../../../theme/colors';
+import COLORS from '../../../../../theme/colors';
 
-export const baseTheme = EditorView.theme(
+export const lightEditorTheme = EditorView.theme(
   {
     '&': {
-      color: Colors.PRIMARY_CONTR_TEXT,
       'min-width': '100%',
       'min-height': '100%',
     },
@@ -15,22 +15,23 @@ export const baseTheme = EditorView.theme(
       'min-height': '100%',
     },
     '.cm-scroller': {
-      background: `${Colors.PRIMARY_DARK}`,
+      background: `${Colors.LIGHT_GREY}`,
     },
     '.cm-gutters': {
-      backgroundColor: '#000',
-      color: '#000',
+      backgroundColor: '#b4b4b4',
+      color: Colors.SECONDARY_DARK,
       border: 'none',
     },
     '.cm-gutter': {},
     '.cm-gutterElement': {
-      color: Colors.DARK_GREY,
+      color: Colors.ACTIVE_TAB,
     },
     '.cm-activeLineGutter': {
       color: Colors.SECONDARY_DARK,
+      backgroundColor: Colors.SECONDARY_MAIN,
     },
     '.cm-content': {
-      caretColor: '#0e9',
+      caretColor: '#b4b4b4',
     },
     '.cm-line': {},
     '&.cm-focused .cm-cursor': {
@@ -38,15 +39,15 @@ export const baseTheme = EditorView.theme(
     },
     '.cm-selectionLayer': {},
     '&.cm-focused .cm-selectionBackground, ::selection': {
-      backgroundColor: '#074',
+      backgroundColor: 'white',
     },
     '.cm-cursorLayer': {},
     '.cm-cursor': {},
   },
-  { dark: true }
+  { dark: false }
 );
 
-export const myHighlightStyle = HighlightStyle.define([
-  { tag: tags.brace, color: Colors.SECONDARY_DARK },
-  { tag: tags.bracket, color: Colors.SECONDARY_DARK },
+export const myHighlightStyleL = HighlightStyle.define([
+  { tag: tags.brace, color: COLORS.SECONDARY_DARK },
+  { tag: tags.bracket, color: COLORS.SECONDARY_DARK },
 ]);
